backend_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
requests==2.32.5
beautifulsoup4==4.14.2
googlesearch-python==1.3.0
pillow==12.0.0
pydantic-settings==2.3.4
python-dotenv==1.2.1